requirements.txt
pandas
fastapi
uvicorn
groq
langchain
sentence-transformers
qdrant-client
python-dotenv
streamlit
